Extract a component for sign typed data info panel
